+++ packages/playground/src/app/components/ComponentFormSandbox.js
@@ -8,6 +8,7 @@ import ComponentForm from '@talend/react-containers/lib/ComponentForm';
 import SidePanel from '@talend/react-containers/lib/SidePanel';
 import HeaderBar from '@talend/react-containers/lib/HeaderBar';
 import { Alert } from '@talend/react-bootstrap';
+// eslint-disable-next-line @talend/import-depth
 import theme from '../example.scss';
 
 const example = require('../../../mockBackend/mock/kit/example.json');