+++ packages/storybook-docs/.storybook/preview.js
@@ -1,0 +1,9 @@
+import '../dist/globalStyles.min.css';
+
+export const parameters = {
+	options: {
+		storySort: {
+			order: ['Components', '*'],
+		},
+	},
+};