+++ tools/scripts-config-react-webpack/config/webpack.config.umd.js
@@ -12,6 +12,7 @@ const {
 	getSassLoaders,
 	getJSAndTSLoader,
 	getSassData,
+	getAssetsRules,
 } = require('./webpack.config.common');
 
 cdn.configureTalendModules();
@@ -61,29 +62,7 @@ module.exports = options => {
 						test: /\.css$/,
 						use: getCommonStyleLoaders(false, options.mode),
 					},
-					{
-						test: /\.svg$/,
-						loader: 'url-loader',
-						options: {
-							esModule: false,
-							name: 'assets/svg/[name].[ext]',
-							mimetype: 'image/svg+xml',
-						},
-					},
-					{
-						test: /\.woff(2)?(\?v=\d+\.\d+\.\d+)?$/,
-						type: 'asset/resource',
-						use: [
-							{
-								loader: 'url-loader',
-								options: {
-									name: './fonts/[name].[ext]',
-									limit: 10000,
-									mimetype: 'application/font-woff',
-								},
-							},
-						],
-					},
+					...getAssetsRules(false),
 				],
 			},
 			stats: { children: false }, // remove warnings of all plugins ...